cleanup

+++ models/SupervisorGroup.class.php
@@ -32,23 +32,25 @@ class SupervisorGroup extends SimpleORMap
     
     public function addUser($user_id)
     {
-        $user                      = new SupervisorGroupUser();
-        $user->supervisor_group_id = $this->id;
-        $user->user_id             = $user_id;
-        $user->store();
+        $user = SupervisorGroupUser::build([
+                'supervisor_group_id' => $this->id,
+                'user_id'             => $user_id
+            ]
+        );
         
-        //als user in alle ePortfolios der StudentInnen eintragen
-        $group    = $this->eportfolio_group;
-        $seminare = $group->getRelatedStudentPortfolios();
-        
-        if ($seminare) {
-            foreach ($seminare as $seminar) {
-                $seminar = new Seminar($seminar);
-                $seminar->addMember($user_id, 'dozent');
-                $seminar->store();
+        if ($user->store()) {
+            //als user in alle ePortfolios der StudentInnen eintragen
+            $group    = $this->eportfolio_group;
+            $seminare = $group->getRelatedStudentPortfolios();
+            
+            if ($seminare) {
+                foreach ($seminare as $seminar) {
+                    $seminar = Seminar::GetInstance($seminar);
+                    $seminar->addMember($user_id, 'dozent');
+                    $seminar->store();
+                }
             }
         }
-        
     }
     
     public function deleteUser($user_id)
@@ -56,14 +58,16 @@ class SupervisorGroup extends SimpleORMap
         //aus Supervisorgruppe austragen
         $user = SupervisorGroupUser::findOneBySQL('user_id = :user_id AND supervisor_group_id = :supervisor_group_id',
             [':user_id' => $user_id, ':supervisor_group_id' => $this->id]);
-        $user->delete();
-        //als user aus allen ePortfolios der StudentInnen austragen
-        $group    = $this->eportfolio_group;
-        $seminare = $group->getRelatedStudentPortfolios();
-        foreach ($seminare as $seminar) {
-            $seminar = new Seminar($seminar);
-            $seminar->deleteMember($user_id);
-            $seminar->store();
+        
+        if ($user->delete()) {
+            //als user aus allen ePortfolios der StudentInnen austragen
+            $group    = $this->eportfolio_group;
+            $seminare = $group->getRelatedStudentPortfolios();
+            foreach ($seminare as $seminar) {
+                $seminar = Seminar::GetInstance($seminar);
+                $seminar->deleteMember($user_id);
+                $seminar->store();
+            }
         }
     }
     
@@ -71,7 +75,7 @@ class SupervisorGroup extends SimpleORMap
     {
         $group       = new SupervisorGroup();
         $group->name = $name;
-        $group->save();
+        $group->store();
     }
     
     public static function deleteGroup($group_id)
